requirements.txt
fastapi
uvicorn
python-multipart
pillow
ultralytics
opencv-python-headless
